refactor : add domain layer
