+++ src/test/java/automaticLotto/domain/WinnerStaticsTest.java
@@ -1,4 +1,4 @@
-package automaticLotto;
+package automaticLotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
